# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Borrow System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  genre: string;
  totalCopies: number;
  availableCopies: number;
  description: string;
}

export interface Borrow {
  id: number;
  bookId: number;
  borrowerName: string;
  borrowedAt: string;
  returnedAt: string | null;
  book?: Book;
}

export interface BorrowBookBody {
  bookId: number;
  borrowerName: string;
}

export interface LibraryStats {
  totalBooks: number;
  totalCopies: number;
  availableCopies: number;
  borrowedCopies: number;
  activeBorrows: number;
}

export interface ErrorResponse {
  error: string;
}