# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Library Borrow System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /borrows:
    get:
      operationId: listBorrows
      tags: [borrows]
      summary: List all borrows
      description: Returns all active borrows
      responses:
        "200":
          description: List of borrows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Borrow"
    post:
      operationId: borrowBook
      tags: [borrows]
      summary: Borrow a book
      description: Creates a borrow record. Fails if no copies available.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowBookBody"
      responses:
        "201":
          description: Book borrowed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrow"
        "400":
          description: No copies available or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /borrows/{id}/return:
    post:
      operationId: returnBook
      tags: [borrows]
      summary: Return a borrowed book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrow"
        "400":
          description: Borrow already returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Borrow not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        description:
          type: string
      required:
        - id
        - title
        - author
        - genre
        - totalCopies
        - availableCopies
        - description
    Borrow:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        borrowerName:
          type: string
        borrowedAt:
          type: string
          format: date-time
        returnedAt:
          type: string
          format: date-time
          nullable: true
        book:
          $ref: "#/components/schemas/Book"
      required:
        - id
        - bookId
        - borrowerName
        - borrowedAt
        - returnedAt
    BorrowBookBody:
      type: object
      properties:
        bookId:
          type: integer
        borrowerName:
          type: string
      required:
        - bookId
        - borrowerName
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error