# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Borrow System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all books with available copy count
 * @summary List all books
 */
export const ListBooksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string(),
  genre: zod.string(),
  totalCopies: zod.number(),
  availableCopies: zod.number(),
  description: zod.string(),
});
export const ListBooksResponse = zod.array(ListBooksResponseItem);

/**
 * @summary Get book by ID
 */
export const GetBookParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string(),
  genre: zod.string(),
  totalCopies: zod.number(),
  availableCopies: zod.number(),
  description: zod.string(),
});

/**
 * Returns all active borrows
 * @summary List all borrows
 */
export const ListBorrowsResponseItem = zod.object({
  id: zod.number(),
  bookId: zod.number(),
  borrowerName: zod.string(),
  borrowedAt: zod.coerce.date(),
  returnedAt: zod.coerce.date().nullable(),
  book: zod
    .object({
      id: zod.number(),
      title: zod.string(),
      author: zod.string(),
      genre: zod.string(),
      totalCopies: zod.number(),
      availableCopies: zod.number(),
      description: zod.string(),
    })
    .optional(),
});
export const ListBorrowsResponse = zod.array(ListBorrowsResponseItem);

/**
 * Creates a borrow record. Fails if no copies available.
 * @summary Borrow a book
 */
export const BorrowBookBody = zod.object({
  bookId: zod.number(),
  borrowerName: zod.string(),
});

/**
 * @summary Return a borrowed book
 */
export const ReturnBookParams = zod.object({
  id: zod.coerce.number(),
});

export const ReturnBookResponse = zod.object({
  id: zod.number(),
  bookId: zod.number(),
  borrowerName: zod.string(),
  borrowedAt: zod.coerce.date(),
  returnedAt: zod.coerce.date().nullable(),
  book: zod
    .object({
      id: zod.number(),
      title: zod.string(),
      author: zod.string(),
      genre: zod.string(),
      totalCopies: zod.number(),
      availableCopies: zod.number(),
      description: zod.string(),
    })
    .optional(),
});

/**
 * Returns aggregate stats about the library
 * @summary Get library stats
 */
export const GetBookStatsResponse = zod.object({
  totalBooks: zod.number(),
  totalCopies: zod.number(),
  availableCopies: zod.number(),
  borrowedCopies: zod.number(),
  activeBorrows: zod.number(),
});