# lib/api-zod/src/generated/types/borrow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Borrow System API
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from "./book";

export interface Borrow {
  id: number;
  bookId: number;
  borrowerName: string;
  borrowedAt: Date;
  returnedAt: Date | null;
  book?: Book;
}