# lib/api-zod/src/generated/types/borrowBookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Borrow System API
 * OpenAPI spec version: 0.1.0
 */

export interface BorrowBookBody {
  bookId: number;
  borrowerName: string;
}